Split up ast_if.rs into distinct rule files
