+++ crates/ruff_linter/src/rules/flake8_simplify/rules/mod.rs
@@ -1,10 +1,15 @@
 pub(crate) use ast_bool_op::*;
 pub(crate) use ast_expr::*;
-pub(crate) use ast_if::*;
 pub(crate) use ast_ifexp::*;
 pub(crate) use ast_unary_op::*;
 pub(crate) use ast_with::*;
+pub(crate) use collapsible_if::*;
+pub(crate) use if_else_block_instead_of_dict_get::*;
+pub(crate) use if_else_block_instead_of_dict_lookup::*;
+pub(crate) use if_else_block_instead_of_if_exp::*;
+pub(crate) use if_with_same_arms::*;
 pub(crate) use key_in_dict::*;
+pub(crate) use needless_bool::*;
 pub(crate) use open_file_with_context_handler::*;
 pub(crate) use reimplemented_builtin::*;
 pub(crate) use return_in_try_except_finally::*;
@@ -13,13 +18,17 @@ pub(crate) use yoda_conditions::*;
 
 mod ast_bool_op;
 mod ast_expr;
-mod ast_if;
 mod ast_ifexp;
 mod ast_unary_op;
 mod ast_with;
-mod fix_if;
+mod collapsible_if;
 mod fix_with;
+mod if_else_block_instead_of_dict_get;
+mod if_else_block_instead_of_dict_lookup;
+mod if_else_block_instead_of_if_exp;
+mod if_with_same_arms;
 mod key_in_dict;
+mod needless_bool;
 mod open_file_with_context_handler;
 mod reimplemented_builtin;
 mod return_in_try_except_finally;
